refactor: move driver types and helpers
